separate the api module and log module
