+++ backend/settings.gradle
@@ -8,4 +8,6 @@ plugins {
 }
 rootProject.name = 'yigil'
 include 'yigil-api'
+include 'support:log'
+findProject(':support:log')?.name = 'log'
