+++ backend/support/log/src/main/java/kr/co/yigil/config/P6SpyConfig.java
@@ -1,4 +1,4 @@
-package kr.co.yigil.global.config;
+package kr.co.yigil.config;
 
 import com.p6spy.engine.logging.Category;
 import com.p6spy.engine.spy.P6SpyOptions;
