+++ backend/support/log/src/main/java/kr/co/yigil/interceptor/PreparedStatementProxyHandler.java
@@ -1,9 +1,9 @@
-package kr.co.yigil.global.log.interceptor;
+package kr.co.yigil.interceptor;
 
 import jakarta.annotation.Nullable;
 import java.lang.reflect.Method;
 import java.util.List;
-import kr.co.yigil.global.log.LoggingForm;
+import kr.co.yigil.LoggingForm;
 import lombok.RequiredArgsConstructor;
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
@@ -35,4 +35,5 @@ public class PreparedStatementProxyHandler implements MethodInterceptor {
 
         return invocation.proceed();
     }
+
 }
