+++ backend/support/src/main/java/kr/co/yigil/Main.java
@@ -1,0 +1,8 @@
+package kr.co.yigil;
+
+public class Main {
+
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
