+++ backend/yigil-api/src/main/java/kr/co/yigil/test/testController.java
@@ -5,12 +5,14 @@ import kr.co.yigil.member.domain.Member;
 import kr.co.yigil.member.domain.SocialLoginType;
 import kr.co.yigil.member.domain.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class testController {
 
     private final MemberRepository repository;
@@ -18,6 +20,7 @@ public class testController {
     @PostMapping("/test")
     public ResponseEntity<String> loginTest(HttpSession session) {
         session.setAttribute("memberId", 1L);
+        log.error("test-error");
         return ResponseEntity.ok("로그인 성공");
     }
 }
